+++ source_gen/lib/source_gen.dart
@@ -2,13 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library source_gen;
-
-export 'src/builder.dart';
+export 'src/builder.dart'
+    show defaultFileHeader, LibraryBuilder, PartBuilder, SharedPartBuilder;
 export 'src/constants/reader.dart' show ConstantReader;
 export 'src/constants/revive.dart' show Revivable;
-export 'src/generator.dart';
-export 'src/generator_for_annotation.dart';
+export 'src/generator.dart' show Generator, InvalidGenerationSourceError;
+export 'src/generator_for_annotation.dart' show GeneratorForAnnotation;
 export 'src/library.dart' show AnnotatedElement, LibraryReader;
 export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker, UnresolvedAnnotationException;